requirements.txt
Flask==0.12.2
numpy==1.11.3
python-dateutil==2.6.0
pytz==2016.10
